feat: embed chat ui in the task sidebar
