+++ site/src/pages/TaskPage/constants.ts
@@ -1,0 +1,1 @@
+export const AI_APP_CHAT_SLUG = "claude-code-web";
